Get Molecule

Loads the structure of a molecule by fetching an SDF file from Pubchem, which can be piped to generate_full_scene

get_molecule(molecule)

Arguments

molecule
A character variable of a compound name or a numeric variable of an official compound ID

Value

List giving the atom locations and the connections between atoms.

Examples

# \donttest{
get_molecule("caffeine")
#> $atoms
#>         x       y z type index
#> 1  3.7320  2.0000 0    O     1
#> 2  2.0000 -1.0000 0    O     2
#> 3  3.7320 -1.0000 0    N     3
#> 4  5.5443  0.8047 0    N     4
#> 5  2.8660  0.5000 0    N     5
#> 6  5.5443 -0.8047 0    N     6
#> 7  4.5981  0.5000 0    C     7
#> 8  4.5981 -0.5000 0    C     8
#> 9  3.7320  1.0000 0    C     9
#> 10 2.8660 -0.5000 0    C    10
#> 11 6.1279  0.0000 0    C    11
#> 12 3.7320 -2.0000 0    C    12
#> 13 5.8550  1.7552 0    C    13
#> 14 2.0000  1.0000 0    C    14
#> 15 6.7479  0.0000 0    H    15
#> 16 3.1120 -2.0000 0    H    16
#> 17 3.7320 -2.6200 0    H    17
#> 18 4.3520 -2.0000 0    H    18
#> 19 6.4443  1.5626 0    H    19
#> 20 6.0476  2.3446 0    H    20
#> 21 5.2656  1.9479 0    H    21
#> 22 2.3100  1.5369 0    H    22
#> 23 1.4631  1.3100 0    H    23
#> 24 1.6900  0.4631 0    H    24
#> 
#> $bonds
#>    from to number
#> 1     1  9      2
#> 2     2 10      2
#> 3     3  8      1
#> 4     3 10      1
#> 5     3 12      1
#> 6     4  7      1
#> 7     4 11      1
#> 8     4 13      1
#> 9     5  9      1
#> 10    5 10      1
#> 11    5 14      1
#> 12    6  8      1
#> 13    6 11      2
#> 14    7  8      2
#> 15    7  9      1
#> 16   11 15      1
#> 17   12 16      1
#> 18   12 17      1
#> 19   12 18      1
#> 20   13 19      1
#> 21   13 20      1
#> 22   13 21      1
#> 23   14 22      1
#> 24   14 23      1
#> 25   14 24      1
#> 
get_molecule(5757) #estradiol (aka estrogen)
#> $atoms
#>         x       y z type index
#> 1  9.1857  2.1128 0    O     1
#> 2  2.5357 -2.1948 0    O     2
#> 3  7.9288  0.8575 0    C     3
#> 4  7.9288 -0.1425 0    C     4
#> 5  7.0628 -0.6425 0    C     5
#> 6  6.1968 -0.1425 0    C     6
#> 7  8.8750  1.1623 0    C     7
#> 8  7.0628  1.3575 0    C     8
#> 9  8.8750 -0.4472 0    C     9
#> 10 6.1968  0.8575 0    C    10
#> 11 9.4586  0.3575 0    C    11
#> 12 7.0789 -1.6840 0    C    12
#> 13 7.9288  1.8575 0    C    13
#> 14 5.2868 -0.6493 0    C    14
#> 15 6.1808 -2.2117 0    C    15
#> 16 5.2787 -1.6909 0    C    16
#> 17 4.3599 -0.0852 0    C    17
#> 18 4.3433 -2.2406 0    C    18
#> 19 3.4120 -0.6132 0    C    19
#> 20 3.4037 -1.6981 0    C    20
#> 21 8.0188 -0.9877 0    H    21
#> 22 7.8022 -1.0618 0    H    22
#> 23 6.9329  0.2825 0    H    23
#> 24 9.4875  1.2585 0    H    24
#> 25 7.4613  1.8325 0    H    25
#> 26 6.6643  1.8325 0    H    26
#> 27 8.6240 -1.0141 0    H    27
#> 28 9.4124 -0.7564 0    H    28
#> 29 5.9847  1.4401 0    H    29
#> 30 5.5862  0.7499 0    H    30
#> 31 9.9195 -0.0572 0    H    31
#> 32 9.9195  0.7723 0    H    32
#> 33 7.2969 -2.2644 0    H    33
#> 34 7.6880 -1.5685 0    H    34
#> 35 8.5488  1.8575 0    H    35
#> 36 7.9288  2.4775 0    H    36
#> 37 7.3088  1.8575 0    H    37
#> 38 5.7835 -2.6877 0    H    38
#> 39 6.5818 -2.6846 0    H    39
#> 40 9.7924  2.2406 0    H    40
#> 41 4.3719  0.5347 0    H    41
#> 42 4.3457 -2.8606 0    H    42
#> 43 2.8787 -0.2970 0    H    43
#> 44 2.0000 -1.8827 0    H    44
#> 
#> $bonds
#>    from to number
#> 1     7  1      1
#> 2     1 40      1
#> 3     2 20      1
#> 4     2 44      1
#> 5     3  4      1
#> 6     3  7      1
#> 7     3  8      1
#> 8     3 13      1
#> 9     4  5      1
#> 10    4  9      1
#> 11    4 21      1
#> 12    5  6      1
#> 13    5 12      1
#> 14    5 22      1
#> 15    6 10      1
#> 16    6 14      1
#> 17    6 23      1
#> 18    7 11      1
#> 19    7 24      1
#> 20    8 10      1
#> 21    8 25      1
#> 22    8 26      1
#> 23    9 11      1
#> 24    9 27      1
#> 25    9 28      1
#> 26   10 29      1
#> 27   10 30      1
#> 28   11 31      1
#> 29   11 32      1
#> 30   12 15      1
#> 31   12 33      1
#> 32   12 34      1
#> 33   13 35      1
#> 34   13 36      1
#> 35   13 37      1
#> 36   14 16      2
#> 37   14 17      1
#> 38   15 16      1
#> 39   15 38      1
#> 40   15 39      1
#> 41   16 18      1
#> 42   17 19      2
#> 43   17 41      1
#> 44   18 20      2
#> 45   18 42      1
#> 46   19 20      1
#> 47   19 43      1
#> 
get_molecule("testosterone")
#> $atoms
#>         x       y z type index
#> 1  8.6500  2.1128 0    O     1
#> 2  2.0000 -2.1948 0    O     2
#> 3  7.3931 -0.1425 0    C     3
#> 4  7.3931  0.8575 0    C     4
#> 5  6.5271 -0.6425 0    C     5
#> 6  5.6610 -0.1425 0    C     6
#> 7  4.7510 -0.6493 0    C     7
#> 8  6.5271  1.3575 0    C     8
#> 9  5.6610  0.8575 0    C     9
#> 10 8.3393  1.1623 0    C    10
#> 11 8.3393 -0.4472 0    C    11
#> 12 6.5431 -1.6840 0    C    12
#> 13 8.9229  0.3575 0    C    13
#> 14 4.7430 -1.6909 0    C    14
#> 15 5.6451 -2.2117 0    C    15
#> 16 7.3931  1.8575 0    C    16
#> 17 3.8242 -0.0852 0    C    17
#> 18 4.7587  0.3507 0    C    18
#> 19 2.8763 -0.6132 0    C    19
#> 20 3.8076 -2.2406 0    C    20
#> 21 2.8680 -1.6981 0    C    21
#> 22 7.4830 -0.9877 0    H    22
#> 23 7.2664 -1.0618 0    H    23
#> 24 6.3972  0.2825 0    H    24
#> 25 6.9256  1.8325 0    H    25
#> 26 6.1285  1.8325 0    H    26
#> 27 5.4490  1.4401 0    H    27
#> 28 5.0504  0.7499 0    H    28
#> 29 7.9018  1.6016 0    H    29
#> 30 8.0883 -1.0141 0    H    30
#> 31 8.8767 -0.7564 0    H    31
#> 32 6.7612 -2.2644 0    H    32
#> 33 7.1523 -1.5685 0    H    33
#> 34 9.3838 -0.0572 0    H    34
#> 35 9.3838  0.7723 0    H    35
#> 36 5.2478 -2.6877 0    H    36
#> 37 6.0460 -2.6846 0    H    37
#> 38 8.0131  1.8575 0    H    38
#> 39 7.3931  2.4775 0    H    39
#> 40 6.7731  1.8575 0    H    40
#> 41 4.2324  0.3814 0    H    41
#> 42 3.4343  0.3968 0    H    42
#> 43 5.3787  0.3459 0    H    43
#> 44 4.7635  0.9706 0    H    44
#> 45 4.1388  0.3554 0    H    45
#> 46 2.6718 -0.0279 0    H    46
#> 47 2.2647 -0.7146 0    H    47
#> 48 3.8100 -2.8606 0    H    48
#> 49 9.2566  2.2406 0    H    49
#> 
#> $bonds
#>    from to number
#> 1    10  1      1
#> 2     1 49      1
#> 3     2 21      2
#> 4     3  4      1
#> 5     3  5      1
#> 6     3 11      1
#> 7     3 22      1
#> 8     4  8      1
#> 9     4 10      1
#> 10    4 16      1
#> 11    5  6      1
#> 12    5 12      1
#> 13    5 23      1
#> 14    6  7      1
#> 15    6  9      1
#> 16    6 24      1
#> 17    7 14      1
#> 18    7 17      1
#> 19    7 18      1
#> 20    8  9      1
#> 21    8 25      1
#> 22    8 26      1
#> 23    9 27      1
#> 24    9 28      1
#> 25   10 13      1
#> 26   10 29      1
#> 27   11 13      1
#> 28   11 30      1
#> 29   11 31      1
#> 30   12 15      1
#> 31   12 32      1
#> 32   12 33      1
#> 33   13 34      1
#> 34   13 35      1
#> 35   14 15      1
#> 36   14 20      2
#> 37   15 36      1
#> 38   15 37      1
#> 39   16 38      1
#> 40   16 39      1
#> 41   16 40      1
#> 42   17 19      1
#> 43   17 41      1
#> 44   17 42      1
#> 45   18 43      1
#> 46   18 44      1
#> 47   18 45      1
#> 48   19 21      1
#> 49   19 46      1
#> 50   19 47      1
#> 51   20 21      1
#> 52   20 48      1
#> 
get_molecule("aspirin")
#> $atoms
#>         x       y z type index
#> 1  3.7320 -0.0600 0    O     1
#> 2  6.3301  1.4400 0    O     2
#> 3  4.5981  1.4400 0    O     3
#> 4  2.8660 -1.5600 0    O     4
#> 5  4.5981 -0.5600 0    C     5
#> 6  5.4641 -0.0600 0    C     6
#> 7  4.5981 -1.5600 0    C     7
#> 8  6.3301 -0.5600 0    C     8
#> 9  5.4641 -2.0600 0    C     9
#> 10 6.3301 -1.5600 0    C    10
#> 11 5.4641  0.9400 0    C    11
#> 12 2.8660 -0.5600 0    C    12
#> 13 2.0000 -0.0600 0    C    13
#> 14 4.0611 -1.8700 0    H    14
#> 15 6.8671 -0.2500 0    H    15
#> 16 5.4641 -2.6800 0    H    16
#> 17 6.8671 -1.8700 0    H    17
#> 18 2.3100  0.4769 0    H    18
#> 19 1.4631  0.2500 0    H    19
#> 20 1.6900 -0.5969 0    H    20
#> 21 6.3301  2.0600 0    H    21
#> 
#> $bonds
#>    from to number
#> 1     1  5      1
#> 2     1 12      1
#> 3     2 11      1
#> 4     2 21      1
#> 5     3 11      2
#> 6     4 12      2
#> 7     5  6      1
#> 8     5  7      2
#> 9     6  8      2
#> 10    6 11      1
#> 11    7  9      1
#> 12    7 14      1
#> 13    8 10      1
#> 14    8 15      1
#> 15    9 10      2
#> 16    9 16      1
#> 17   10 17      1
#> 18   12 13      1
#> 19   13 18      1
#> 20   13 19      1
#> 21   13 20      1
#> 
# }